fix(ci): line-limit guard - split oversized cleanup edge tests
